+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
  * `ContractCreateFlow.executeWithSigner()`
+ * `PublicKey.fromString()` for ECDSA keys
 
 ## v2.18.2